@@ -8,6 +8,7 @@ module ShopifyApp
       source_root File.expand_path("../templates", __FILE__)
 
       def generate_app_extension
+        ShopifyApp::Logger.deprecated("MarketingActivitiesController will be removed in an upcoming version", "22.0.0")
         template("marketing_activities_controller.rb", "app/controllers/marketing_activities_controller.rb")
         generate_routes
       end
@@ -18,7 +19,7 @@ module ShopifyApp
         inject_into_file(
           "config/routes.rb",
           optimize_indentation(routes, 2),
-          after: "root :to => 'home#index'\n"
+          after: "root :to => 'home#index'\n",
         )
       end
 
