@@ -5,12 +5,10 @@ module ShopifyApp
     extend ActiveSupport::Concern
 
     included do
-      include ShopifyApp::Localization
-      include ShopifyApp::LoginProtection
-      include ShopifyApp::CsrfProtection
-      include ShopifyApp::EmbeddedApp
-      before_action :login_again_if_different_user_or_shop
-      around_action :activate_shopify_session
+      ShopifyApp::Logger.deprecated("Authenticated has been replaced by EnsureHasSession."\
+        " Please use the EnsureHasSession controller concern for the same behavior", "22.0.0")
     end
+
+    include ShopifyApp::EnsureHasSession
   end
 end
