@@ -14,7 +14,8 @@ module ShopifyApp
       splash_page_with_params(
         return_to: request.fullpath,
         shop: current_shopify_domain,
-        host: params[:host]
+        host: params[:host],
+        embedded: params[:embedded],
       )
     end
 
@@ -26,9 +27,9 @@ module ShopifyApp
 
     def redirect_to_splash_page
       redirect_to(splash_page)
-    rescue ShopifyApp::LoginProtection::ShopifyDomainNotFound => error
-      Rails.logger.warn("[ShopifyApp::EnsureAuthenticatedLinks] Redirecting to login: [#{error.class}] "\
-        "Could not determine current shop domain")
+    rescue ::ShopifyApp::ShopifyDomainNotFound => error
+      ShopifyApp::Logger.warn("Redirecting to login: [#{error.class}]"\
+        " Could not determine current shop domain")
       redirect_to(ShopifyApp.configuration.login_url)
     end
 
