@@ -3,38 +3,11 @@
 module ShopifyApp
   module RequireKnownShop
     extend ActiveSupport::Concern
+    include ShopifyApp::EnsureInstalled
 
     included do
-      before_action :check_shop_domain
-      before_action :check_shop_known
-    end
-
-    def current_shopify_domain
-      return if params[:shop].blank?
-      @shopify_domain ||= ShopifyApp::Utils.sanitize_shop_domain(params[:shop])
-    end
-
-    private
-
-    def check_shop_domain
-      redirect_to(ShopifyApp.configuration.login_url) unless current_shopify_domain
-    end
-
-    def check_shop_known
-      @shop = SessionRepository.retrieve_shop_session_by_shopify_domain(current_shopify_domain)
-      redirect_to(shop_login) unless @shop
-    end
-
-    def shop_login
-      url = URI(ShopifyApp.configuration.login_url)
-
-      url.query = URI.encode_www_form(
-        shop: params[:shop],
-        host: params[:host],
-        return_to: request.fullpath,
-      )
-
-      url.to_s
+      ShopifyApp::Logger.deprecated("RequireKnownShop has been replaced by EnsureInstalled."\
+        " Please use the EnsureInstalled controller concern for the same behavior", "22.0.0")
     end
   end
 end
