@@ -3,6 +3,7 @@
 module ShopifyApp
   module ShopAccessScopesVerification
     extend ActiveSupport::Concern
+    include ShopifyApp::RedirectForEmbedded
 
     included do
       before_action :login_on_scope_changes
@@ -11,7 +12,13 @@ module ShopifyApp
     protected
 
     def login_on_scope_changes
-      redirect_to(shop_login) if scopes_mismatch?
+      if scopes_mismatch?
+        if embedded_param?
+          redirect_for_embedded
+        else
+          redirect_to(shop_login)
+        end
+      end
     end
 
     private
@@ -22,6 +29,7 @@ module ShopifyApp
 
     def current_shopify_domain
       return if params[:shop].blank?
+
       ShopifyApp::Utils.sanitize_shop_domain(params[:shop])
     end
 
