@@ -2,7 +2,7 @@
 
 module ShopifyApp
   class AuthenticatedController < ActionController::Base
-    include ShopifyApp::Authenticated
+    include ShopifyApp::EnsureHasSession
 
     protect_from_forgery with: :exception
   end
