@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 class AuthenticatedController < ApplicationController
-  include ShopifyApp::Authenticated
+  include ShopifyApp::EnsureHasSession
 end
