@@ -2,11 +2,15 @@
 
 class HomeController < ApplicationController
   include ShopifyApp::EmbeddedApp
-  include ShopifyApp::RequireKnownShop
+  include ShopifyApp::EnsureInstalled
   include ShopifyApp::ShopAccessScopesVerification
 
   def index
-    @shop_origin = current_shopify_domain
-    @host = params[:host]
+    if ShopifyAPI::Context.embedded? && (!params[:embedded].present? || params[:embedded] != "1")
+      redirect_to(ShopifyAPI::Auth.embedded_app_url(params[:host]) + request.path, allow_other_host: true)
+    else
+      @shop_origin = current_shopify_domain
+      @host = params[:host]
+    end
   end
 end
