@@ -3,13 +3,18 @@
 module ShopifyApp
   module AccessScopes
     class UserStrategy
-      class InvalidInput < StandardError; end
-
       class << self
         def update_access_scopes?(user_id: nil, shopify_user_id: nil)
           return update_access_scopes_for_user_id?(user_id) if user_id
           return update_access_scopes_for_shopify_user_id?(shopify_user_id) if shopify_user_id
-          raise(InvalidInput, "#update_access_scopes? requires user_id or shopify_user_id parameter inputs")
+
+          raise(::ShopifyApp::InvalidInput,
+            "#update_access_scopes? requires user_id or shopify_user_id parameter inputs")
+        end
+
+        def covers_scopes?(current_shopify_session)
+          # NOTE: this not Ruby's `covers?` method, it is defined in ShopifyAPI::Auth::AuthScopes
+          current_shopify_session.scope.to_a.empty? || current_shopify_session.scope.covers?(ShopifyAPI::Context.scope)
         end
 
         private
