@@ -6,7 +6,7 @@ module ShopifyApp
       ShopifyApp.configuration.webhooks_manager_queue_name
     end
 
-    def perform(shop_domain:, shop_token:)
+    def perform(shop_domain:, shop_token:, **)
       ShopifyAPI::Auth::Session.temp(shop: shop_domain, access_token: shop_token) do |session|
         WebhooksManager.create_webhooks(session: session)
       end
