@@ -2,20 +2,19 @@
 
 module ShopifyApp
   class ScripttagsManager
-    class CreationFailed < StandardError; end
-
     def self.queue(shop_domain, shop_token, scripttags)
       ShopifyApp::ScripttagsManagerJob.perform_later(
         shop_domain: shop_domain,
         shop_token: shop_token,
         # Procs cannot be serialized so we interpolate now, if necessary
-        scripttags: build_src(scripttags, shop_domain)
+        scripttags: build_src(scripttags, shop_domain),
       )
     end
 
     def self.build_src(scripttags, domain)
       scripttags.map do |tag|
         next tag unless tag[:src].respond_to?(:call)
+
         tag = tag.dup
         tag[:src] = tag[:src].call(domain)
         tag
@@ -25,6 +24,7 @@ module ShopifyApp
     attr_reader :required_scripttags, :shop_domain
 
     def initialize(scripttags, shop_domain)
+      ShopifyApp::Logger.deprecated("The ScripttagsManager will become deprecated in an upcoming version", "22.0.0")
       @required_scripttags = scripttags
       @shop_domain = shop_domain
     end
@@ -68,7 +68,7 @@ module ShopifyApp
       begin
         scripttag.save!
       rescue ShopifyAPI::Errors::HttpResponseError => e
-        raise CreationFailed, e.message
+        raise ::ShopifyApp::CreationFailed, e.message
       end
 
       scripttag
