@@ -2,20 +2,14 @@
 
 module ShopifyApp
   class JWT
-    class InvalidDestinationError < StandardError; end
-
-    class MismatchedHostsError < StandardError; end
-
-    class InvalidAudienceError < StandardError; end
-
     WARN_EXCEPTIONS = [
       ::JWT::DecodeError,
       ::JWT::ExpiredSignature,
       ::JWT::ImmatureSignature,
       ::JWT::VerificationError,
-      InvalidAudienceError,
-      InvalidDestinationError,
-      MismatchedHostsError,
+      ::ShopifyApp::InvalidAudienceError,
+      ::ShopifyApp::InvalidDestinationError,
+      ::ShopifyApp::MismatchedHostsError,
     ]
 
     def initialize(token)
@@ -40,8 +34,7 @@ module ShopifyApp
     def set_payload
       payload, _ = parse_token_data(ShopifyApp.configuration&.secret, ShopifyApp.configuration&.old_secret)
       @payload = validate_payload(payload)
-    rescue *WARN_EXCEPTIONS => error
-      Rails.logger.warn("[ShopifyApp::JWT] Failed to validate JWT: [#{error.class}] #{error}")
+    rescue *WARN_EXCEPTIONS
       nil
     end
 
@@ -58,9 +51,12 @@ module ShopifyApp
       iss_host = ShopifyApp::Utils.sanitize_shop_domain(payload["iss"])
       api_key = ShopifyApp.configuration.api_key
 
-      raise InvalidAudienceError, "'aud' claim does not match api_key" unless payload["aud"] == api_key
-      raise InvalidDestinationError, "'dest' claim host not a valid shopify host" unless dest_host
-      raise MismatchedHostsError, "'dest' claim host does not match 'iss' claim host" unless dest_host == iss_host
+      raise ::ShopifyApp::InvalidAudienceError,
+        "'aud' claim does not match api_key" unless payload["aud"] == api_key
+      raise ::ShopifyApp::InvalidDestinationError, "'dest' claim host not a valid shopify host" unless dest_host
+
+      raise ::ShopifyApp::MismatchedHostsError,
+        "'dest' claim host does not match 'iss' claim host" unless dest_host == iss_host
 
       payload
     end
