@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module ShopifyApp
-  VERSION = "19.0.0-alpha.1"
+  VERSION = "21.4.1"
 end
