@@ -6,6 +6,6 @@ ShopifyApp.configure do |config|
   config.scope = "read_orders, read_products"
   config.embedded_app = true
   config.webhooks = [
-    { topic: "carts/update", address: "https://example.com/webhooks/carts_update" },
+    { topic: "carts/update", path: "webhooks/carts_update" },
   ]
 end
