@@ -61,8 +61,8 @@ class EnsureAuthenticatedLinksTest < ActionController::TestCase
   end
 
   test "redirects to login page if current shopify domain is not found" do
-    @controller.expects(:current_shopify_domain).raises(ShopifyApp::LoginProtection::ShopifyDomainNotFound)
-    expect_redirect_error(ShopifyApp::LoginProtection::ShopifyDomainNotFound, "Could not determine current shop domain")
+    @controller.expects(:current_shopify_domain).raises(::ShopifyApp::ShopifyDomainNotFound)
+    expect_redirect_error(ShopifyApp::ShopifyDomainNotFound, "Could not determine current shop domain")
 
     get :some_link
 
@@ -72,7 +72,7 @@ class EnsureAuthenticatedLinksTest < ActionController::TestCase
   private
 
   def expect_redirect_error(klass, message)
-    expected_message = "[ShopifyApp::EnsureAuthenticatedLinks] Redirecting to login: [#{klass}] #{message}"
-    Rails.logger.expects(:warn).once.with(expected_message)
+    expected_message = "Redirecting to login: [#{klass}] #{message}"
+    ShopifyApp::Logger.expects(:warn).once.with(expected_message)
   end
 end
