@@ -12,8 +12,10 @@ module ShopifyApp
   class SessionsControllerTest < ActionController::TestCase
     setup do
       @routes = ShopifyApp::Engine.routes
+      ShopifyApp.configuration.api_version = ShopifyAPI::LATEST_SUPPORTED_ADMIN_VERSION
       ShopifyApp::SessionRepository.shop_storage = ShopifyApp::InMemoryShopSessionStore
       ShopifyApp::SessionRepository.user_storage = nil
+      ShopifyAppConfigurer.setup_context  # need to reset context after config changes
 
       I18n.locale = :en
 
@@ -76,6 +78,13 @@ module ShopifyApp
       assert_redirected_to_top_level(shopify_domain)
     end
 
+    test "#new redirects to the embedded url if a valid shop param exists and embedded param exists" do
+      ShopifyApp.configuration.embedded_redirect_url = "/a-redirect-page"
+      shopify_domain = "my-shop.myshopify.com"
+      get :new, params: { shop: "my-shop", embedded: 1 }
+      assert_redirected_to_embedded(shopify_domain, ShopifyApp.configuration.embedded_redirect_url)
+    end
+
     test "#new stores root path when return_to url is absolute" do
       get :new, params: { shop: "my-shop", return_to: "//example.com" }
       assert_equal "/", session[:return_to]
@@ -97,17 +106,36 @@ module ShopifyApp
       assert_redirected_to "/auth-route"
     end
 
-    test "#new starts OAuth requesting online token if user session is expected" do
+    test "#new starts OAuth requesting offline token if user session is expected but there is no shop session" do
       ShopifyApp::SessionRepository.user_storage = ShopifyApp::InMemoryUserSessionStore
 
       ShopifyAPI::Auth::Oauth.expects(:begin_auth)
-        .with(shop: "my-shop.myshopify.com", redirect_path: "/auth/shopify/callback", is_online: true)
+        .with(shop: "my-shop.myshopify.com", redirect_path: "/auth/shopify/callback", is_online: false)
         .returns({
           cookie: ShopifyAPI::Auth::Oauth::SessionCookie.new(value: "", expires: Time.now),
           auth_route: "/auth-route",
         })
 
       get :new, params: { shop: "my-shop", top_level: true }
+    end
+
+    test "#new starts OAuth requesting online token if user session is expected and there is a shop session" do
+      shop = "my-shop.myshopify.com"
+
+      ShopifyApp::SessionRepository.user_storage = ShopifyApp::InMemoryUserSessionStore
+      ShopifyApp::SessionRepository.shop_storage = ShopifyApp::InMemoryShopSessionStore
+      ShopifyApp::SessionRepository.shop_storage.stubs(:retrieve_by_shopify_domain)
+        .with(shop)
+        .returns(mock_session(shop: shop))
+
+      ShopifyAPI::Auth::Oauth.expects(:begin_auth)
+        .with(shop: shop, redirect_path: "/auth/shopify/callback", is_online: true)
+        .returns({
+          cookie: ShopifyAPI::Auth::Oauth::SessionCookie.new(value: "", expires: Time.now),
+          auth_route: "/auth-route",
+        })
+
+      get :new, params: { shop: shop, top_level: true }
     end
 
     test "#new starts OAuth requesting online token if user session is unexpected" do
@@ -158,8 +186,12 @@ module ShopifyApp
       refute session[:user_tokens]
     end
 
-    ["my-shop", "my-shop.myshopify.com", "https://my-shop.myshopify.com",
-     "http://my-shop.myshopify.com",].each do |good_url|
+    [
+      "my-shop",
+      "my-shop.myshopify.com",
+      "https://my-shop.myshopify.com",
+      "http://my-shop.myshopify.com",
+    ].each do |good_url|
       test "#create should authenticate the shop for the URL (#{good_url})" do
         shopify_domain = "my-shop.myshopify.com"
         post :create, params: { shop: good_url }
@@ -167,8 +199,12 @@ module ShopifyApp
       end
     end
 
-    ["my-shop", "my-shop.myshopify.io", "https://my-shop.myshopify.io",
-     "http://my-shop.myshopify.io",].each do |good_url|
+    [
+      "my-shop",
+      "my-shop.myshopify.io",
+      "https://my-shop.myshopify.io",
+      "http://my-shop.myshopify.io",
+    ].each do |good_url|
       test "#create should authenticate the shop for the URL (#{good_url}) with custom myshopify_domain" do
         ShopifyApp.configuration.myshopify_domain = "myshopify.io"
         shopify_domain = "my-shop.myshopify.io"
@@ -177,10 +213,118 @@ module ShopifyApp
       end
     end
 
+    [
+      "my-shop",
+      "my-shop.myshopify.com",
+      "https://my-shop.myshopify.com",
+      "http://my-shop.myshopify.com",
+    ].each do |good_url|
+      test "#create should authenticate the shop for the URL (#{good_url}) with embedded param" do
+        ShopifyApp.configuration.embedded_redirect_url = "/a-redirect-page"
+        shopify_domain = "my-shop.myshopify.com"
+        post :create, params: { shop: good_url, embedded: 1 }
+        assert_redirected_to_embedded(shopify_domain, ShopifyApp.configuration.embedded_redirect_url)
+      end
+    end
+
+    [
+      "my-shop",
+      "my-shop.myshopify.io",
+      "https://my-shop.myshopify.io",
+      "http://my-shop.myshopify.io",
+    ].each do |good_url|
+      test "#create should authenticate the shop for the URL (#{good_url}) with custom myshopify_domain with embedded param" do
+        ShopifyApp.configuration.embedded_redirect_url = "/a-redirect-page"
+        ShopifyApp.configuration.myshopify_domain = "myshopify.io"
+        shopify_domain = "my-shop.myshopify.io"
+        post :create, params: { shop: good_url, embedded: 1 }
+        assert_redirected_to_embedded(shopify_domain, ShopifyApp.configuration.embedded_redirect_url)
+      end
+    end
+
+    [
+      "my-shop",
+      "my-shop.myshopify.com",
+      "https://my-shop.myshopify.com",
+      "http://my-shop.myshopify.com",
+    ].each do |good_url|
+      test "#create should redirect to auth route when embedded_redirect_url configured but no embedded param for the URL (#{good_url})" do
+        ShopifyApp.configuration.embedded_redirect_url = "/a-redirect-page"
+        ShopifyAPI::Auth::Oauth.stubs(:begin_auth).returns({
+          cookie: ShopifyAPI::Auth::Oauth::SessionCookie.new(value: "", expires: Time.now),
+          auth_route: "/auth-route",
+        })
+        post :create, params: { shop: good_url }
+        assert_redirected_to "/auth-route"
+      end
+    end
+
+    [
+      "my-shop",
+      "my-shop.myshopify.io",
+      "https://my-shop.myshopify.io",
+      "http://my-shop.myshopify.io",
+    ].each do |good_url|
+      test "#create should redirect to toplevel when embedded_redirect_url configured but no embedded param for the URL (#{good_url}) with custom myshopify_domain" do
+        ShopifyApp.configuration.embedded_redirect_url = "/a-redirect-page"
+        ShopifyApp.configuration.myshopify_domain = "myshopify.io"
+        ShopifyAPI::Auth::Oauth.stubs(:begin_auth).returns({
+          cookie: ShopifyAPI::Auth::Oauth::SessionCookie.new(value: "", expires: Time.now),
+          auth_route: "/auth-route",
+        })
+        post :create, params: { shop: good_url }
+        assert_redirected_to "/auth-route"
+      end
+    end
+
+    [
+      "my-shop",
+      "my-shop.myshopify.com",
+      "https://my-shop.myshopify.com",
+      "http://my-shop.myshopify.com",
+    ].each do |good_url|
+      test "#create should redirect to toplevel when embedded_redirect_url is not configured but embedded param sent for the URL (#{good_url})" do
+        shopify_domain = "my-shop.myshopify.com"
+        post :create, params: { shop: good_url, embedded: 1 }
+        assert_redirected_to_top_level(shopify_domain)
+      end
+    end
+
+    [
+      "my-shop",
+      "my-shop.myshopify.io",
+      "https://my-shop.myshopify.io",
+      "http://my-shop.myshopify.io",
+    ].each do |good_url|
+      test "#create should redirect to toplevel when embedded_redirect_url is not configured but embedded param sent for the URL (#{good_url}) with custom myshopify_domain" do
+        ShopifyApp.configuration.myshopify_domain = "myshopify.io"
+        shopify_domain = "my-shop.myshopify.io"
+        post :create, params: { shop: good_url, embedded: 1 }
+        assert_redirected_to_top_level(shopify_domain)
+      end
+    end
+
     ["myshop.com", "myshopify.com", "shopify.com", "two words",
      "store.myshopify.com.evil.com", "/foo/bar",].each do |bad_url|
       test "#create should return an error for a non-myshopify URL (#{bad_url})" do
         post :create, params: { shop: bad_url }
+        assert_response :redirect
+        assert_redirected_to "/"
+        assert_equal I18n.t("invalid_shop_url"), flash[:error]
+      end
+    end
+
+    [
+      "myshop.com",
+      "myshopify.com",
+      "shopify.com",
+      "two words",
+      "store.myshopify.com.evil.com",
+      "/foo/bar",
+    ].each do |bad_url|
+      test "#create should return an error for a non-myshopify URL (#{bad_url}) with embedded param" do
+        ShopifyApp.configuration.embedded_redirect_url = "/a-redirect-page"
+        post :create, params: { shop: bad_url, embedded: 1 }
         assert_response :redirect
         assert_redirected_to "/"
         assert_equal I18n.t("invalid_shop_url"), flash[:error]
@@ -234,6 +378,14 @@ module ShopifyApp
         assert_equal "{\"myshopifyUrl\":\"https://#{shop_domain}\",\"url\":\"#{expected_url}\"}",
           elements.first["data-target"]
       end
+    end
+
+    def assert_redirected_to_embedded(shop_domain, base_embedded_url = nil)
+      assert_not_nil base_embedded_url
+      redirect_uri = "https://test.host/login?shop=#{shop_domain}"
+      expected_url = base_embedded_url + "?embedded=1&redirectUri=#{CGI.escape(redirect_uri)}" + "&shop=#{shop_domain}"
+
+      assert_redirected_to(expected_url)
     end
   end
 end
