@@ -32,11 +32,11 @@ class InstallGeneratorTest < Rails::Generators::TestCase
               api_key: ShopifyApp.configuration.api_key,
               api_secret_key: ShopifyApp.configuration.secret,
               api_version: ShopifyApp.configuration.api_version,
-              host_name: URI(ENV.fetch('HOST', '')).host || '',
+              host: ENV['HOST'],
               scope: ShopifyApp.configuration.scope,
               is_private: !ENV.fetch('SHOPIFY_APP_PRIVATE_SHOP', '').empty?,
               is_embedded: ShopifyApp.configuration.embedded_app,
-              session_storage: ShopifyApp::SessionRepository,
+              log_level: :info,
               logger: Rails.logger,
               private_shop: ENV.fetch('SHOPIFY_APP_PRIVATE_SHOP', nil),
               user_agent_prefix: "ShopifyApp/\#{ShopifyApp::VERSION}"
@@ -97,7 +97,10 @@ class InstallGeneratorTest < Rails::Generators::TestCase
   test "adds host config to development.rb" do
     run_generator
     assert_file "config/environments/development.rb" do |config|
-      assert_match "config.hosts = (config.hosts rescue []) << /\[-\\w]+\\.ngrok\\.io/", config
+      assert_match "Allow ngrok tunnels for secure Shopify OAuth redirects\n", config
+      assert_match "config.hosts = (config.hosts rescue []) << /\[-\\w]+\\.ngrok\\.io/\n", config
+      assert_match "Allow Cloudflare tunnels for secure Shopify OAuth redirects\n", config
+      assert_match "config.hosts = (config.hosts rescue []) << /\[-\\w]+\\.trycloudflare\\.com/\n", config
     end
   end
 end
