@@ -32,11 +32,13 @@ class LoginProtectionController < ActionController::Base
   end
 
   def raise_unauthorized
-    raise ShopifyAPI::Errors::HttpResponseError.new(code: 401), "unauthorized"
+    unauthorized_response = ShopifyAPI::Clients::HttpResponse.new(code: 401, headers: {}, body: "")
+    raise ShopifyAPI::Errors::HttpResponseError.new(response: unauthorized_response), "unauthorized"
   end
 
   def raise_not_found
-    raise ShopifyAPI::Errors::HttpResponseError.new(code: 404), "not found"
+    not_found_response = ShopifyAPI::Clients::HttpResponse.new(code: 404, headers: {}, body: "")
+    raise ShopifyAPI::Errors::HttpResponseError.new(response: not_found_response), "not found"
   end
 end
 
@@ -44,49 +46,135 @@ class LoginProtectionControllerTest < ActionController::TestCase
   tests LoginProtectionController
 
   setup do
+    @shop = "my-shop.myshopify.com"
+
     ShopifyApp::SessionRepository.shop_storage = ShopifyApp::InMemoryShopSessionStore
     ShopifyApp::SessionRepository.user_storage = ShopifyApp::InMemoryUserSessionStore
+
+    @token = "Bearer Grylls da token"
+    @session = ShopifyAPI::Auth::Session.new(id: "1", shop: @shop)
+    ShopifyApp::SessionRepository.store_shop_session(@session)
+
+    ShopifyApp.configuration.old_secret = "old_secret"
+    ShopifyAPI::Context.stubs(:activate_session)
+    ShopifyApp::SessionRepository.stubs(:load_session)
+      .returns(mock_session(shop: @shop))
+    ::ShopifyAPI::Utils::SessionUtils.stubs(:current_session_id).returns(@session.id)
+    ::ShopifyAPI::Context.stubs(:activate_session)
+
+    ShopifyAPI::Context.setup(
+      api_key: ShopifyApp.configuration.api_key,
+      api_secret_key: ShopifyApp.configuration.secret,
+      old_api_secret_key: ShopifyApp.configuration.old_secret,
+      api_version: ShopifyAPI::LATEST_SUPPORTED_ADMIN_VERSION,
+      host_name: "host.example.io",
+      scope: ShopifyApp.configuration.scope,
+      session_storage: ShopifyApp::SessionRepository,
+      is_private: false,
+      is_embedded: true,
+    )
 
     request.env["HTTP_USER_AGENT"] = "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_13_6) "\
       "AppleWebKit/537.36 (KHTML, like Gecko) Chrome/69.0.3497.100 Safari/537.36"
   end
 
+  test "force reauthentication when reauth_on_access_scope_changes enabled" do
+    ShopifyApp.configuration.scope = "read_retail_private_data"
+    ShopifyApp.configuration.reauth_on_access_scope_changes = true
+
+    ShopifyAPI::Context.setup(
+      api_key: ShopifyApp.configuration.api_key,
+      api_secret_key: ShopifyApp.configuration.secret,
+      old_api_secret_key: ShopifyApp.configuration.old_secret,
+      api_version: ShopifyAPI::LATEST_SUPPORTED_ADMIN_VERSION,
+      host_name: "host.example.io",
+      scope: ShopifyApp.configuration.scope,
+      session_storage: ShopifyApp::SessionRepository,
+      is_private: false,
+      is_embedded: true,
+    )
+
+    with_application_test_routes do
+      request.headers["HTTP_AUTHORIZATION"] = @token
+
+      ::ShopifyAPI::Utils::SessionUtils.expects(:current_session_id)
+        .with(
+          @token,
+          { ShopifyAPI::Auth::Oauth::SessionCookie::SESSION_COOKIE_NAME => nil },
+          true,
+        ).returns(@session.id)
+      ::ShopifyAPI::Context.expects(:activate_session).never
+
+      get :index, params: { shop: @shop }
+    end
+  end
+
+  test "don't force reauthentication when reauth_on_access_scope_changes disabled" do
+    ShopifyApp.configuration.scope = "read_retail_private_data"
+    ShopifyApp.configuration.reauth_on_access_scope_changes = false
+
+    ShopifyAPI::Context.setup(
+      api_key: ShopifyApp.configuration.api_key,
+      api_secret_key: ShopifyApp.configuration.secret,
+      old_api_secret_key: ShopifyApp.configuration.old_secret,
+      api_version: ShopifyAPI::LATEST_SUPPORTED_ADMIN_VERSION,
+      host_name: "host.example.io",
+      scope: ShopifyApp.configuration.scope,
+      session_storage: ShopifyApp::SessionRepository,
+      is_private: false,
+      is_embedded: true,
+    )
+
+    with_application_test_routes do
+      request.headers["HTTP_AUTHORIZATION"] = @token
+
+      ::ShopifyAPI::Utils::SessionUtils.expects(:current_session_id)
+        .with(
+          @token,
+          { ShopifyAPI::Auth::Oauth::SessionCookie::SESSION_COOKIE_NAME => nil },
+          true,
+        ).returns(@session.id)
+      ::ShopifyAPI::Context.expects(:activate_session)
+
+      get :index, params: { shop: @shop }
+    end
+  end
+
   test "#current_shopify_session returns nil when session is nil" do
     with_application_test_routes do
+      ::ShopifyAPI::Utils::SessionUtils.stubs(:current_session_id).raises(ShopifyAPI::Errors::CookieNotFoundError)
+
       session[ShopifyAPI::Auth::Oauth::SessionCookie::SESSION_COOKIE_NAME] = nil
-      get :index
       assert_nil @controller.current_shopify_session
     end
   end
 
   test "#current_shopify_session loads online session if user session expected" do
-    cookies.encrypted[ShopifyAPI::Auth::Oauth::SessionCookie::SESSION_COOKIE_NAME] = "cookie"
-    request.headers["HTTP_AUTHORIZATION"] = "Bearer token"
-
-    ShopifyAPI::Utils::SessionUtils.expects(:load_current_session)
-      .with(
-        auth_header: "Bearer token",
-        cookies: { ShopifyAPI::Auth::Oauth::SessionCookie::SESSION_COOKIE_NAME => "cookie" },
-        is_online: true
-      )
-      .returns(nil)
-
     with_application_test_routes do
-      get :index
+      request.headers["HTTP_AUTHORIZATION"] = @token
+
+      ::ShopifyAPI::Utils::SessionUtils.expects(:current_session_id)
+        .with(
+          @token,
+          { ShopifyAPI::Auth::Oauth::SessionCookie::SESSION_COOKIE_NAME => nil },
+          true,
+        ).returns(@session.id)
+      ::ShopifyAPI::Context.expects(:activate_session)
+
+      get :index, params: { shop: @shop }
     end
   end
 
   test "#current_shopify_session loads offline session if user session unexpected" do
     ShopifyApp::SessionRepository.user_storage = nil
 
-    cookies.encrypted[ShopifyAPI::Auth::Oauth::SessionCookie::SESSION_COOKIE_NAME] = "cookie"
-    request.headers["HTTP_AUTHORIZATION"] = "Bearer token"
+    request.headers["HTTP_AUTHORIZATION"] = @token
 
-    ShopifyAPI::Utils::SessionUtils.expects(:load_current_session)
+    ShopifyAPI::Utils::SessionUtils.expects(:current_session_id)
       .with(
-        auth_header: "Bearer token",
-        cookies: { ShopifyAPI::Auth::Oauth::SessionCookie::SESSION_COOKIE_NAME => "cookie" },
-        is_online: false
+        @token,
+        { ShopifyAPI::Auth::Oauth::SessionCookie::SESSION_COOKIE_NAME => nil },
+        false,
       )
       .returns(nil)
 
@@ -96,23 +184,19 @@ class LoginProtectionControllerTest < ActionController::TestCase
   end
 
   test "#current_shopify_session is nil if token is invalid" do
-    cookies.encrypted[ShopifyAPI::Auth::Oauth::SessionCookie::SESSION_COOKIE_NAME] = "cookie"
     request.headers["HTTP_AUTHORIZATION"] = "Bearer invalid"
 
     with_application_test_routes do
+      ::ShopifyAPI::Utils::SessionUtils.stubs(:current_session_id).returns(nil)
       get :index
       assert_nil @controller.current_shopify_session
     end
   end
 
   test "#current_shopify_session is memoized and does not retrieve session twice" do
-    shop_session_record = ShopifyAPI::Auth::Session.new(
-      shop: "my-shop",
-      access_token: "1234",
-    )
     with_application_test_routes do
+      ShopifyApp::SessionRepository.expects(:load_session).returns(@session).once
       get :index
-      ShopifyAPI::Utils::SessionUtils.expects(:load_current_session).returns(shop_session_record).once
       assert @controller.current_shopify_session
     end
   end
@@ -120,7 +204,7 @@ class LoginProtectionControllerTest < ActionController::TestCase
   test "#login_again_if_different_user_or_shop removes current cookie if the session changes" do
     cookies.encrypted[ShopifyAPI::Auth::Oauth::SessionCookie::SESSION_COOKIE_NAME] = "cookie"
 
-    ShopifyAPI::Utils::SessionUtils.stubs(:load_current_session)
+    ShopifyAPI::Utils::SessionUtils.stubs(:current_session_id)
       .returns(ShopifyAPI::Auth::Session.new(shop: "shop", shopify_session_id: "123"))
 
     with_application_test_routes do
@@ -134,8 +218,8 @@ class LoginProtectionControllerTest < ActionController::TestCase
     with_application_test_routes do
       cookies.encrypted[ShopifyAPI::Auth::Oauth::SessionCookie::SESSION_COOKIE_NAME] = "old-cookie"
 
-      ShopifyAPI::Utils::SessionUtils.expects(:load_current_session).returns(
-        ShopifyAPI::Auth::Session.new(shop: "some-shop")
+      ShopifyAPI::Utils::SessionUtils.expects(:current_session_id).returns(
+        ShopifyAPI::Auth::Session.new(shop: "some-shop"),
       ).once
 
       get :second_login
@@ -148,8 +232,8 @@ class LoginProtectionControllerTest < ActionController::TestCase
     with_application_test_routes do
       cookies.encrypted[ShopifyAPI::Auth::Oauth::SessionCookie::SESSION_COOKIE_NAME] = "old-cookie"
 
-      ShopifyAPI::Utils::SessionUtils.expects(:load_current_session).returns(
-        ShopifyAPI::Auth::Session.new(shop: "some-shop")
+      ShopifyAPI::Utils::SessionUtils.expects(:current_session_id).returns(
+        ShopifyAPI::Auth::Session.new(shop: "some-shop"),
       ).once
 
       get :second_login, params: { shop: "other-shop" }
@@ -161,47 +245,45 @@ class LoginProtectionControllerTest < ActionController::TestCase
   end
 
   test "#current_shopify_session redirects to login if the loaded session doesn't have enough scope" do
+    ShopifyApp::SessionRepository.shop_storage.stubs(:retrieve_by_shopify_domain)
+      .with(@shop)
+      .returns(mock_session(shop: @shop))
     ShopifyAPI::Context.stubs(:scope).returns(ShopifyAPI::Auth::AuthScopes.new(["scope1", "scope2"]))
 
     cookies.encrypted[ShopifyAPI::Auth::Oauth::SessionCookie::SESSION_COOKIE_NAME] = "cookie"
     request.headers["HTTP_AUTHORIZATION"] = "Bearer token"
 
-    ShopifyAPI::Utils::SessionUtils.stubs(:load_current_session)
+    ShopifyAPI::Utils::SessionUtils.stubs(:current_session_id)
       .with(
-        auth_header: "Bearer token",
-        cookies: { ShopifyAPI::Auth::Oauth::SessionCookie::SESSION_COOKIE_NAME => "cookie" },
-        is_online: true
+        "Bearer token",
+        { ShopifyAPI::Auth::Oauth::SessionCookie::SESSION_COOKIE_NAME => "cookie" },
+        true,
       )
       .returns(
-        ShopifyAPI::Auth::Session.new(shop: "some-shop", scope: ["scope1"])
+        ShopifyAPI::Auth::Session.new(shop: @shop, scope: ["scope1"]),
       )
 
     with_application_test_routes do
-      get :index
+      get :index, params: { shop: @shop }
 
-      assert_redirected_to "/login"
+      assert_redirected_to "/login?shop=my-shop.myshopify.com"
       assert_nil cookies.encrypted[ShopifyAPI::Auth::Oauth::SessionCookie::SESSION_COOKIE_NAME]
     end
   end
 
   test "#current_shopify_session does not redirect when sufficient scope" do
-    ShopifyAPI::Context.stubs(:scope).returns(ShopifyAPI::Auth::AuthScopes.new(["scope1"]))
-
-    cookies.encrypted[ShopifyAPI::Auth::Oauth::SessionCookie::SESSION_COOKIE_NAME] = "cookie"
-    request.headers["HTTP_AUTHORIZATION"] = "Bearer token"
-
-    ShopifyAPI::Utils::SessionUtils.stubs(:load_current_session)
-      .with(
-        auth_header: "Bearer token",
-        cookies: { ShopifyAPI::Auth::Oauth::SessionCookie::SESSION_COOKIE_NAME => "cookie" },
-        is_online: true
-      )
-      .returns(
-        ShopifyAPI::Auth::Session.new(shop: "some-shop", scope: ["scope1", "scope2"])
-      )
-
     with_application_test_routes do
-      get :index
+      ShopifyApp::SessionRepository.stubs(:load_session).returns(mock_session(shop: @shop))
+      ShopifyAPI::Context.stubs(:scope).returns(ShopifyAPI::Auth::AuthScopes.new(["scope1"]))
+
+      cookies.encrypted[ShopifyAPI::Auth::Oauth::SessionCookie::SESSION_COOKIE_NAME] = "cookie"
+      request.headers["HTTP_AUTHORIZATION"] = @token
+
+      ShopifyApp::SessionRepository.expects(:load_session).returns(
+        ShopifyAPI::Auth::Session.new(shop: "some-shop", scope: ["scope1", "scope2"]),
+      )
+
+      get :index, params: { shop: @shop }
       assert_response :ok
       assert_equal "cookie", cookies.encrypted[ShopifyAPI::Auth::Oauth::SessionCookie::SESSION_COOKIE_NAME]
     end
@@ -211,8 +293,8 @@ class LoginProtectionControllerTest < ActionController::TestCase
     with_application_test_routes do
       cookies.encrypted[ShopifyAPI::Auth::Oauth::SessionCookie::SESSION_COOKIE_NAME] = "old-cookie"
 
-      ShopifyAPI::Utils::SessionUtils.expects(:load_current_session).returns(
-        ShopifyAPI::Auth::Session.new(shop: "some-shop")
+      ShopifyAPI::Utils::SessionUtils.expects(:current_session_id).returns(
+        ShopifyAPI::Auth::Session.new(shop: "some-shop"),
       ).once
 
       get :second_login, params: { shop: { id: 123, disabled: true } }
@@ -222,6 +304,7 @@ class LoginProtectionControllerTest < ActionController::TestCase
 
   test "#activate_shopify_session with no Shopify session, redirects to the login url" do
     with_application_test_routes do
+      ::ShopifyAPI::Utils::SessionUtils.stubs(:current_session_id).returns(nil)
       get :index, params: { shop: "foobar" }
       assert_redirected_to "/login?shop=foobar.myshopify.com"
     end
@@ -230,6 +313,7 @@ class LoginProtectionControllerTest < ActionController::TestCase
   test "#activate_shopify_session with no Shopify session, redirects to a custom config login url" do
     with_custom_login_url "https://domain.com/custom/route/login" do
       with_application_test_routes do
+        ::ShopifyAPI::Utils::SessionUtils.stubs(:current_session_id).returns(nil)
         get :index, params: { shop: "foobar" }
         assert_redirected_to "https://domain.com/custom/route/login?shop=foobar.myshopify.com"
       end
@@ -240,6 +324,7 @@ class LoginProtectionControllerTest < ActionController::TestCase
         shop param of referer" do
     with_application_test_routes do
       ShopifyApp.configuration.user_session_repository = nil
+      ::ShopifyAPI::Utils::SessionUtils.stubs(:current_session_id).returns(nil)
       @controller.expects(:current_shopify_session).returns(nil)
       request.headers["Referer"] = "https://example.com/?shop=my-shop.myshopify.com"
 
@@ -253,6 +338,7 @@ class LoginProtectionControllerTest < ActionController::TestCase
     with_custom_login_url "https://domain.com/custom/route/login" do
       with_application_test_routes do
         ShopifyApp.configuration.user_session_repository = nil
+        ::ShopifyAPI::Utils::SessionUtils.stubs(:current_session_id).returns(nil)
         @controller.expects(:current_shopify_session).returns(nil)
         request.headers["Referer"] = "https://example.com/?shop=my-shop.myshopify.com"
 
@@ -265,6 +351,7 @@ class LoginProtectionControllerTest < ActionController::TestCase
   test '#activate_shopify_session with no Shopify session, redirects to the login url \
         with non-String shop param' do
     with_application_test_routes do
+      ::ShopifyAPI::Utils::SessionUtils.stubs(:current_session_id).returns(nil)
       params = { shop: { id: 123 } }
       get :index, params: params
       assert_redirected_to "/login?#{params.to_query}"
@@ -274,6 +361,7 @@ class LoginProtectionControllerTest < ActionController::TestCase
   test '#activate_shopify_session with no Shopify session, redirects to a custom config login url \
         with non-String shop param' do
     with_custom_login_url "https://domain.com/custom/route/login" do
+      ::ShopifyAPI::Utils::SessionUtils.stubs(:current_session_id).returns(nil)
       with_application_test_routes do
         params = { shop: { id: 123 } }
         get :index, params: params
@@ -284,6 +372,7 @@ class LoginProtectionControllerTest < ActionController::TestCase
 
   test "#activate_shopify_session with no Shopify session, sets session[:return_to]" do
     with_application_test_routes do
+      ::ShopifyAPI::Utils::SessionUtils.stubs(:current_session_id).returns(nil)
       get :index, params: { shop: "foobar" }
       assert_equal "/?shop=foobar.myshopify.com", session[:return_to]
     end
@@ -292,6 +381,7 @@ class LoginProtectionControllerTest < ActionController::TestCase
   test '#activate_shopify_session with no Shopify session, sets session[:return_to]\
         with non-String shop param' do
     with_application_test_routes do
+      ::ShopifyAPI::Utils::SessionUtils.stubs(:current_session_id).returns(nil)
       params = { shop: { id: 123 } }
       get :index, params: params
       assert_equal "/?#{params.to_query}", session[:return_to]
@@ -300,6 +390,7 @@ class LoginProtectionControllerTest < ActionController::TestCase
 
   test "#activate_shopify_session with no Shopify session, when the request is a POST, sets session[:return_to]" do
     with_application_test_routes do
+      ::ShopifyAPI::Utils::SessionUtils.stubs(:current_session_id).returns(nil)
       request.headers["Referer"] = "https://example.com/?id=123"
       post :index, params: { id: "123", shop: "foobar" }
       assert_equal "/?id=123&shop=foobar.myshopify.com", session[:return_to]
@@ -308,8 +399,11 @@ class LoginProtectionControllerTest < ActionController::TestCase
 
   test "#activate_shopify_session with no Shopify session, when the request is an XHR, returns an HTTP 401" do
     with_application_test_routes do
+      ::ShopifyAPI::Utils::SessionUtils.stubs(:current_session_id).returns(nil)
       get :index, params: { shop: "foobar" }, xhr: true
       assert_equal 401, response.status
+      assert_match "1", response.headers["X-Shopify-API-Request-Failure-Reauthorize"]
+      assert_match "/login?shop=foobar", response.headers["X-Shopify-API-Request-Failure-Reauthorize-Url"]
     end
   end
 
@@ -346,7 +440,7 @@ class LoginProtectionControllerTest < ActionController::TestCase
   test "#fullpage_redirect_to, when the shop params is missing, sends a post message to the shop in the session" do
     with_application_test_routes do
       example_shop = "shop.myshopify.com"
-      ShopifyAPI::Utils::SessionUtils.expects(:load_current_session)
+      ShopifyApp::SessionRepository.expects(:load_session)
         .returns(ShopifyAPI::Auth::Session.new(shop: example_shop))
       get :redirect
       assert_fullpage_redirected(example_shop, response)
@@ -355,9 +449,9 @@ class LoginProtectionControllerTest < ActionController::TestCase
 
   test "#fullpage_redirect_to raises an exception when no Shopify domains are available" do
     with_application_test_routes do
-      ShopifyAPI::Utils::SessionUtils.expects(:load_current_session)
+      ShopifyAPI::Utils::SessionUtils.expects(:current_session_id)
         .returns(nil)
-      assert_raise ShopifyApp::LoginProtection::ShopifyDomainNotFound do
+      assert_raise ::ShopifyApp::ShopifyDomainNotFound do
         get :redirect
       end
     end
@@ -404,6 +498,20 @@ class LoginProtectionControllerTest < ActionController::TestCase
 
       assert_equal expire_at - 5.seconds, @controller.jwt_expire_at
     end
+  end
+
+  test "detects incompatible controller concerns" do
+    version = "22.0.0"
+
+    ShopifyApp::Logger.expects(:deprecated).with(regexp_matches(/incompatible concerns/), version)
+    ShopifyApp::Logger.stubs(:deprecated).with("Itp will be removed in an upcoming version", version)
+
+    Class.new(ApplicationController) do
+      include ShopifyApp::LoginProtection
+      include ShopifyApp::EnsureInstalled
+    end
+
+    assert_within_deprecation_schedule(version)
   end
 
   private
