@@ -5,8 +5,13 @@ module SessionStoreStrategyTestHelpers
     attr_reader :id, :shopify_domain, :shopify_token, :api_version, :access_scopes
     attr_writer :shopify_token, :access_scopes
 
-    def initialize(id: 1, shopify_domain: "example.myshopify.com",
-      shopify_token: "abcd-shop-token", api_version: "unstable", scopes: "read_products")
+    def initialize(
+      id: 1,
+      shopify_domain: "example.myshopify.com",
+      shopify_token: "abcd-shop-token",
+      api_version: ShopifyApp.configuration.api_version,
+      scopes: "read_products"
+    )
       @id = id
       @shopify_domain = shopify_domain
       @shopify_token = shopify_token
@@ -19,8 +24,14 @@ module SessionStoreStrategyTestHelpers
     attr_reader :id, :shopify_user_id, :shopify_domain, :shopify_token, :api_version, :access_scopes
     attr_writer :shopify_token, :shopify_domain, :access_scopes
 
-    def initialize(id: 1, shopify_user_id: 1, shopify_domain: "example.myshopify.com",
-      shopify_token: "1234-user-token", api_version: "unstable", scopes: "read_products")
+    def initialize(
+      id: 1,
+      shopify_user_id: 1,
+      shopify_domain: "example.myshopify.com",
+      shopify_token: "1234-user-token",
+      api_version: ShopifyApp.configuration.api_version,
+      scopes: "read_products"
+    )
       @id = id
       @shopify_user_id = shopify_user_id
       @shopify_domain = shopify_domain
